Refactor Lattice-Boltzmann module: rename `state` to `velocity` and refactor equilibrium logic
- Renamed `state` property to `velocity` for better clarity and alignment with grid velocity context.
- Updated `LBMD2Q9State` field, methods, and related documentation to reflect the new naming convention.
- Introduced `sum_dims` in the `compat` module for compatibility with upcoming Burn API changes.
- Refactored `UCellPartials` and `UCellTerms` to `EquilibriumPartials` and `EquilibriumTerms` for consistent naming across the module.
- Enhanced `ucell_partials` to support tensor broadcasting and alignment with canonicalized dimensions.
- Adjusted benchmarks, examples, and tests to align with these changes.

+++ crates/clockmill/benches/lbm.rs
@@ -22,7 +22,7 @@ fn bench_lbm(c: &mut Criterion) {
             Tensor::<B, 4>::random([n, n, 3, 3], Distribution::Normal(0., 1.), &device).cast(dtype);
         let solid_mask = Tensor::<B, 2>::zeros([n, n], &device).bool();
 
-        let ops = LBMD2Q9Operations::<B>::init(&device).cast(dtype);
+        let ops = LBMD2Q9Operations::<B>::init(&device).to_dtype(dtype);
 
         group.bench_function(format!("{:?} equilibrium", dtype).as_str(), |b| {
             b.iter(|| {
